Uses 'class_alias' instead of 'extends', move dbLayer to ns
