--- legacy/common/lib.crypt.php
@@ -1,3 +1,0 @@
-<?php
-
-class crypt extends Clearbricks\Common\Crypt { }
