--- legacy/common/lib.date.php
@@ -1,3 +1,0 @@
-<?php
-
-class dt extends Clearbricks\Common\Dt { }
