--- legacy/common/lib.files.php
@@ -1,5 +1,0 @@
-<?php
-
-class files extends Clearbricks\Common\Files { }
-
-class path extends Clearbricks\Common\Path { }
