--- legacy/common/lib.form.php
@@ -1,5 +1,0 @@
-<?php
-
-class form extends Clearbricks\Common\Form { }
-
-class formSelectOption extends Clearbricks\Common\FormSelectOption { }
