--- legacy/common/lib.forms.php
@@ -1,5 +1,0 @@
-<?php
-
-class forms extends Clearbricks\Common\Forms { }
-
-class formsSelectOption extends Clearbricks\Common\FormsSelectOption { }
