--- legacy/common/lib.html.php
@@ -1,3 +1,0 @@
-<?php
-
-class html extends Clearbricks\Common\Html { }
