--- legacy/common/lib.http.php
@@ -1,3 +1,0 @@
-<?php
-
-class http extends Clearbricks\Common\Http { }
