--- legacy/common/lib.l10n.php
@@ -1,3 +1,0 @@
-<?php
-
-class l10n extends Clearbricks\Common\L10n { }
