--- legacy/common/lib.text.php
@@ -1,3 +1,0 @@
-<?php
-
-class text extends Clearbricks\Common\Text { }
